Increase search limit from int to long

+++ SWSH_OWRNG_Generator.WinForms/MainWindow.cs
@@ -260,7 +260,7 @@ namespace SWSH_OWRNG_Generator.WinForms
         private void LevelSlot_TextChanged(object sender, EventArgs e)
         {
             TextBox textBox = (TextBox)sender;
-            if (int.TryParse(textBox.Text, out int i))
+            if (long.TryParse(textBox.Text, out long i))
             {
                 if (i < 0) textBox.Text = "0";
                 if (i > 99) textBox.Text = "99";
